pyproject.toml
[project]
name = "agentdevelopmentkit"
version = "0.1.0"
description = "Multi-Agent Document Analysis Platform with Google ADK and MCP Integration"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    # Core Google ADK and MCP
    "google-adk>=0.4.0",
    
    # API and Web Framework
    "fastapi>=0.115.12",
    "uvicorn>=0.34.2",
    "chainlit>=2.5.5",
    
    # LLM Integration
    "litellm>=1.70.0",
    
    # Document Processing (Required)
    "docling>=2.15.1",
    "docx2txt>=0.9",
    "pdfplumber>=0.11.6",
    "python-docx>=1.1.2",
    
    # NLP and Text Processing (Required)
    "fuzzywuzzy>=0.18.0",
    "spacy>=3.8.4",
    "transformers>=4.51.3",
    
    # Web UI (Required)
    "gradio>=5.33.0",
    
    # Utilities
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]

alternative-ui = [
    "streamlit>=1.45.1",
]

all = [
    "streamlit>=1.45.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.black]
line-length = 88
target-version = ['py312']

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"